+++ modules/title_selection/choose_superset.py
@@ -4,19 +4,22 @@ import itertools
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
+    from modules.config import Config
     from modules.dats import DatNode
 
 from modules.titletools import TraceTools
 from modules.utils import Font
 
 
-def choose_superset(title_set: set[DatNode], report_on_match: bool) -> set[DatNode]:
+def choose_superset(title_set: set[DatNode], config: Config, report_on_match: bool) -> set[DatNode]:
     """
     Compares any two titles from a set of DatNodes, and if one is a superset,
     chooses it.
 
     Args:
         title_set (set[DatNode]): A set of titles as DatNode instances.
+
+        config (Config): The Retool config object.
 
         report_on_match (bool): Whether Retool needs to report any titles being
         traced.
@@ -35,14 +38,27 @@ def choose_superset(title_set: set[DatNode], report_on_match: bool) -> set[DatNo
             and 'BIOS' not in title_1.categories
             and 'BIOS' not in title_2.categories
         ):
+
+            keep_title_name: str = ''
+            remove_title_name: str = ''
+
             if title_1.is_superset and not title_2.is_superset:
                 if title_1 in title_set:
-                    remove_titles.add(title_2)
+                    if config.user_input.oldest:
+                        keep_title_name = title_2.full_name
+                        remove_title_name = title_1.full_name
 
-                    if report_on_match:
+                        remove_titles.add(title_1)
+                    else:
+                        keep_title_name = title_1.full_name
+                        remove_title_name = title_2.full_name
+
+                        remove_titles.add(title_2)
+
+                    if report_on_match and keep_title_name:
                         TraceTools.trace_title(
                             'REF0077',
-                            [f'{title_1.full_name}', f'{title_2.full_name}{Font.end}'],
+                            [f'{keep_title_name}', f'{remove_title_name}{Font.end}'],
                             keep_remove=True,
                         )
 
@@ -50,12 +66,21 @@ def choose_superset(title_set: set[DatNode], report_on_match: bool) -> set[DatNo
 
             if title_2.is_superset and not title_1.is_superset:
                 if title_2 in title_set:
-                    remove_titles.add(title_1)
+                    if config.user_input.oldest:
+                        keep_title_name = title_1.full_name
+                        remove_title_name = title_2.full_name
+
+                        remove_titles.add(title_2)
+                    else:
+                        keep_title_name = title_2.full_name
+                        remove_title_name = title_1.full_name
+
+                        remove_titles.add(title_1)
 
                     if report_on_match:
                         TraceTools.trace_title(
                             'REF0078',
-                            [f'{title_2.full_name}', f'{title_1.full_name}{Font.end}'],
+                            [f'{keep_title_name}', f'{remove_title_name}{Font.end}'],
                             keep_remove=True,
                         )